+++ R/utils.R
@@ -1,3 +1,34 @@
+# get the shiny session object
+getSession <- function() {
+  session <- shiny::getDefaultReactiveDomain()
+
+  if (is.null(session)) {
+    stop("Could not find the Shiny session object. This usually happens if you're trying to call a function outside of a Shiny app.", call. = FALSE)
+  }
+
+  session
+}
+
+getDependencies <- function() {
+  list(
+    htmltools::htmlDependency(
+      name = "shinycssloaders-binding",
+      version = as.character(utils::packageVersion("shinycssloaders")),
+      package = "shinycssloaders",
+      src = "assets",
+      script = "spinner.js",
+      stylesheet = "spinner.css"
+    ),
+    htmltools::htmlDependency(
+      name = "cssloaders",
+      version = as.character(utils::packageVersion("shinycssloaders")),
+      package = "shinycssloaders",
+      src = "assets",
+      stylesheet = "css-loaders.css"
+    )
+  )
+}
+
 get_proxy_element <- function(ui_element, proxy.height, hide.ui) {
   if (!hide.ui) {
     return(shiny::tagList())
@@ -21,26 +52,39 @@ get_proxy_element <- function(ui_element, proxy.height, hide.ui) {
   }
 }
 
-add_style <- function(x) {
-  if (x == "") {
-    return(NULL)
+get_spinner_css_tag <- function(type, color, size, color.background, custom.css, id, image, caption, output_spinner) {
+  base_css <- ""
+  add_default_style <- (is.null(image) && !custom.css && type != 0)
+  if (add_default_style) {
+    if (type %in% c(2, 3) && is.null(color.background)) {
+      stop("shinycssloaders: For spinner types 2 & 3 you need to specify `color.background`.")
+    }
+
+    color.rgb <- paste(grDevices::col2rgb(color), collapse = ",")
+    color.alpha0 <- sprintf("rgba(%s, 0)", color.rgb)
+    color.alpha2 <- sprintf("rgba(%s, 0.2)", color.rgb)
+
+    css_file <- system.file(glue::glue("loaders-templates/load{type}.css"), package="shinycssloaders")
+    if (file.exists(css_file)) {
+      base_css <- paste(readLines(css_file), collapse = " ")
+      base_css <- glue::glue(base_css, .open = "{{", .close = "}}")
+    }
+
+    # get default font-size from css, and cut it by 25%, as for outputs we usually need something smaller
+    size_px <- round(c(11, 11, 10, 20, 25, 90, 10, 10)[type] * size * 0.75)
+    base_css <- paste(base_css, glue::glue("#{id} {{ font-size: {size_px}px; }}"))
   }
-  shiny::tags$head(
-    shiny::tags$style(
-      shiny::HTML(
-        x
-      )
-    )
-  )
-}
 
-# get the shiny session object
-getSession <- function() {
-  session <- shiny::getDefaultReactiveDomain()
-
-  if (is.null(session)) {
-    stop("Could not find the Shiny session object. This usually happens if you're trying to call a function outside of a Shiny app.", call. = FALSE)
+  if (!is.null(caption)) {
+    base_css <- paste(base_css, glue::glue("#{id}__caption {{ color: {color}; }}"))
   }
 
-  session
+  css_rules_tag <- NULL
+  if (nzchar(base_css)) {
+    css_rules_tag <- shiny::tags$head(shiny::tags$style(
+      class = if (!output_spinner) "global-spinner-css",
+      shiny::HTML(base_css)
+    ))
+  }
+  css_rules_tag
 }