+++ NEWS.md
@@ -1,7 +1,7 @@
 # Unreleased version
 
 - Add `showSpinner()` and `hideSpinner()` that allow you to manually trigger the spinner (#56)
-- Add support for full-page spinner that covers the entire page. The spinner needs to be added to the UI with `pageSpinner()` and it can be triggered with `showPageSpinner()` and `hidePageSpinner()`. (#42)
+- Add support for full-page spinner that covers the entire page using `showPageSpinner()` and `hidePageSpinner()` (#42)
 - Add support for setting global options for most parameters (#78)
 - Fix page sometimes wobbling because of hidden spinner (#54)
 - Fix bug where `shiny::tagList()` would throw an error (#73)